refactor: move test to tests and eventtagservice to application layer
