+++ features/calendar/event_tag.py
@@ -12,8 +12,7 @@ from slack_sdk.models.blocks.block_elements import PlainTextInputElement, Static
 from slack_sdk.web import WebClient
 
 from application.event_tag import EventTagData
-from application.event_tag.repository import EventTagRepository
-from infrastructure.api_client import get_api_event_tag_repository
+from application.event_tag.service import EventTagService
 from utilities.builders import add_loading_form
 from utilities.constants import EVENT_TAG_COLORS
 from utilities.database.orm import SlackSettings
@@ -28,36 +27,6 @@ EVENT_TAG_EDIT_DELETE = "event-tag-edit-delete"
 CALENDAR_ADD_EVENT_TAG_CALLBACK_ID = "calendar-add-event-tag-id"
 EDIT_DELETE_AO_CALLBACK_ID = "edit-delete-ao-id"
 CALENDAR_EVENT_TAG_COLORS_IN_USE = "calendar-event-tag-colors-in-use"
-
-
-class EventTagService:
-    """
-    Business logic for event tags.  All data access is delegated to an
-    ``EventTagRepository``; the default implementation uses the F3 Nation API.
-    """
-
-    def __init__(self, repository: EventTagRepository | None = None) -> None:
-        self._repository: EventTagRepository = repository or get_api_event_tag_repository()
-
-    def get_org_event_tags(self, org_id: str) -> list[EventTagData]:
-        """Return org-specific event tags for *org_id*."""
-        return self._repository.get_by_org(int(org_id))
-
-    def get_event_tag_by_id(self, tag_id: int) -> EventTagData | None:
-        """Return a single event tag, or *None* if not found."""
-        return self._repository.get_by_id(tag_id)
-
-    def create_org_specific_tag(self, name: str, color: str, org_id: str) -> None:
-        """Create a new org-specific event tag."""
-        self._repository.create(name, color, int(org_id))
-
-    def update_org_specific_tag(self, tag_id: int, name: str, color: str) -> None:
-        """Update the name and colour of an existing event tag."""
-        self._repository.update(tag_id, name, color)
-
-    def delete_org_specific_tag(self, tag_id: int) -> None:
-        """Soft-delete an event tag."""
-        self._repository.delete(tag_id)
 
 
 class EventTagViews:
